pyproject.toml
[project]
name = "asistente-bi"
version = "0.1.0"
description = "Asistente de BI con Exploración de Datos Interactiva GenBI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "sqlalchemy>=2.0.46",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "pydantic>=2.12.5",
    "pandas>=3.0.0",
    "langchain>=1.2.8",
    "langchain-core>=1.2.8",
    "langchain-google-genai",
    "langchain-community",
    "google-generativeai",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.28.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
